Crate kitties

An NFT game inspired by cryptokitties. This is a game which allows for kitties to be bred based on a few factors 1.) Mom and Tired have to be in a state where they are ready to breed 2.) Each Mom and Dad have some DNA and the child will have unique DNA combined from the both of them Linkable back to the Mom and Dad 3.) The game also allows Kitties to have a cooling off period inbetween breeding before they can be bred again. 4.) A rest operation allows for a Mom Kitty and a Dad Kitty to be cooled off

In order to submit a valid transaction you must strutucture it as follows: 1.) Input must contain 1 mom and 1 dad 2.) Output must contain Mom, Dad, and newly created Child 3.) A child’s DNA is calculated by: BlakeTwo256::hash_of(MomDna, DadDna, MomCurrNumBreedings, DadCurrNumberBreedings)

There are a only a finite amount of free breedings available before it starts to cost money to breed kitties.

Structs

FreeKittyConstraintChecker

KittyDNA

KittyData

KittyHelpers

Enums

ConstraintCheckerError

DadKittyStatus

MomKittyStatus

Parent